requirements.txt
async-generator==1.10
attrs==22.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
flake8==6.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyproject_api==1.5.0
PySocks==1.7.1
pytest==7.2.1
python-dotenv==0.21.1
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.19.0
wsproto==1.2.0
